Fix #156: Use consistent signature in docs for `interpret()`
Also assigned the `interpret(…)` call to a `const service`, to help illustrate the relationship between these two terms.

(I just started reading the docs yesterday.  Some terminology is obvious, like the fact that `createMachine()` returns a “machine”.  But the relationship between calling `interpret()` and a “service” was not immediately obvious.)