Improve the type definition for `state` and `invoke` functions (#260)
* Fix the type definition for `state` and `invoke` functions

`state` accepts multiple transitions, each of which can have its own type.
However, the typescript type was only extracting the type of the first argument.
Update the type of `state` to extract the type of each of its arguments, and
union them together.

Also update the `invoke` function's type to propagate the transitions from a
child state machine. This isn't 100% correct, since it adds the child's
transitions to the parent, but previously it wasn't even possible to call
`send` on the child, so this is probably an improvement.

* Add changeset